requirements.txt
appdirs
pyparsing
async-generator==1.6
asyncio-extras==1.3.0
attrs==16.3.0
-e git+git@github.com:timdiels/chicken_turtle_util.git@d0a8c1c6a6b9b4c25679962a262a433133c7e244#egg=chicken_turtle_util
click==6.7
lxml==3.7.3
numpy==1.12.0
numpydoc==0.6.0
packaging==16.8
patool==1.12
pkginfo==1.4.1
plumbum==1.6.3
psutil==5.2.1
py==1.4.32
Pympler==0.5
pytest==3.0.6
pytest-asyncio==0.5.0
pytest-catchlog==1.2.2
pytest-localserver==0.3.7
pytest-mock==1.6.0
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
